Replace HttpClient with IHttpClientFactory
The HttpClient instances have been replaced with the use of IHttpClientFactory across multiple parts of the application. This change provides benefits such as having a central place for naming and configuring logical HttpClient instances as needed, offering the ability to manage the lifetime of the underlying HttpMessageHandler instances automatically.

Currently storing/editing does not seem to be working for unknown reasons